+++ dms-enterprise-20181101/src/Models/ListDataLakeDatabaseResponseBody.php
@@ -29,6 +29,16 @@ class ListDataLakeDatabaseResponseBody extends Model
     public $errorMessage;
 
     /**
+     * @var int
+     */
+    public $maxResults;
+
+    /**
+     * @var string
+     */
+    public $nextToken;
+
+    /**
      * @example 4E1D2B4D-3E53-4ABC-999D-1D2520B3471A
      *
      * @var string
@@ -45,6 +55,8 @@ class ListDataLakeDatabaseResponseBody extends Model
         'databaseList' => 'DatabaseList',
         'errorCode'    => 'ErrorCode',
         'errorMessage' => 'ErrorMessage',
+        'maxResults'   => 'MaxResults',
+        'nextToken'    => 'NextToken',
         'requestId'    => 'RequestId',
         'success'      => 'Success',
     ];
@@ -64,6 +76,12 @@ class ListDataLakeDatabaseResponseBody extends Model
         }
         if (null !== $this->errorMessage) {
             $res['ErrorMessage'] = $this->errorMessage;
+        }
+        if (null !== $this->maxResults) {
+            $res['MaxResults'] = $this->maxResults;
+        }
+        if (null !== $this->nextToken) {
+            $res['NextToken'] = $this->nextToken;
         }
         if (null !== $this->requestId) {
             $res['RequestId'] = $this->requestId;
@@ -91,6 +109,12 @@ class ListDataLakeDatabaseResponseBody extends Model
         }
         if (isset($map['ErrorMessage'])) {
             $model->errorMessage = $map['ErrorMessage'];
+        }
+        if (isset($map['MaxResults'])) {
+            $model->maxResults = $map['MaxResults'];
+        }
+        if (isset($map['NextToken'])) {
+            $model->nextToken = $map['NextToken'];
         }
         if (isset($map['RequestId'])) {
             $model->requestId = $map['RequestId'];